requirements.txt
google-cloud-bigquery==3.13.0
json-normalize==1.0.1
pandas==2.1.3
pyarrow==14.0.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0